requirements.txt
Flask==2.2.2
torch==1.13.1
torchvision==0.14.1
Pillow==9.3.0
pandas==1.5.3
numpy==1.23.5
matplotlib==3.6.2
seaborn==0.12.2
opencv-python==4.6.0.66
scikit-learn==1.2.0
tqdm==4.64.1
torchinfo==1.7.0
